Add make.scm package definition, fix uses of local-file
I was using the , operator on local-file calls in build definitions.
This sometimes works, but sometimes results in a local-file *struct*
appearing in the builder code (instead of the path to the existing
file). I'm not sure why the behavior is inconsistent, but examining uses
of local-file in gnu/packages shows that the idiom is to add the
local-file struct to the native-inputs field, gexp the modify-phases
form, and ungexp the instance of local-package.