Remove module declaration from package file
Installing the package module doesn't work well due to the local-file
calls being hardcoded, and since you need to use the file to bootstrap
anyway it doesn't seem that useful to place it in a module.

Removing the module declaration makes it more convenient to use in
terminal one-liners, for example `guix shell -De '(begin (load
"package.scm") guile-todo)'.