Use SRFI-88 keyword syntax & use period for namepacing
README was updated with reasoning for SRFI-88 syntax usage, including
the decision to use the #: syntax in some places still.

Using the period for namespacing avoids issues with the namespace
looking like a keyword in some circumstances. In particular, syntax
highlighting can sometimes be confused by this, and it seems
inconsistent between modules whether the `use-module` form sees a
namespace like guix: as a keyword or a symbol. If it is seen as a
keyword then this is an error. The inconsistency probably has to do with
the dependency chain, but I didn't look at it too closely because it's
easier to just avoid the problem, and it doesn't seem like a bug, just a
tangled knot.