Heavy refactoring of test code.
These changes largely came from a motivation to make it easier to write
integration tests for code depending on the serialization module. Some
other cleanup tasks were taken care of as well.

package.scm:
Instead of using the magic function main in test modules, call run-tests
on the magic symbol all-tests in test modules. This reduces boilerplate
in test modules, and provides more uniform expectations for the contents
of the magic symbol.

r7rs/base/test.scm:
Switch to all-tests instead of main, as described in the package.scm
notes.

r7rs/base/test/util.scm:
First, handle exceptions gracefully. They will now be caught and printed
to stdout as well as the log port, while still causing the test to fail.
This improvement means that the full test results will be available
regardless of where the first failure occurs. The printing of the
exception could be improved, however.

Second, instead of providing the run-test function which decorates a
thunk with the test start/end boilerplate, provide a make-test function
which returns a thunk that includes the start/end boilerplate. This
follows the virtuous programming pattern that any macro which defines a
symbol should be paired with an accompanying function or macro that
makes the thing which is defined. This pattern is useful because it
means that the caller has full access to the functionality without the
constraint of needing to define a symbol, or needing to learn different
context requirements for using the helper vs referring to a symbol
defined with the primary macro. It is less cognitive overhead.

test/util.scm:
First, since all of this code was related to the serialization module
anyway, make it a serialization submodule.

Second, change example-data to always have one name/datum pair in each
element instead of a list of data in some elements. This allows for more
descriptive naming, and also simplifies the logic for generating the
tests. Alse rename it from example-data to primitives, since this is
more meaningful.

Third, export make-serialization-test (previously named
test-example-datum) so that dependent modules can use it for easy
integration testing.