Add changes made after un-bungling system
My system got confused with multiple versions of guix running (don't
ask), and while trying to figure out the problem I did a number of
cleanup things to simplify root-causing it. There were also some changes
in the working directory while I was debugging, and these are in this
commit as well.

: Fixing keyword-reading

Loading the postfix read syntax for keywords causes errors when
dependent files using a trailing colon for the prefix of module imports,
which is very common. Currently using an idiom where the read syntax is
reset to default before the define-module form, then set to postfix
after the define-module form. Making this option per-file is on my todo
list, it looks like guile already has some infrastructure in place for
this and it might be already implemented and I just need to set it in a
different way.

: Minimize main system

When I'm in my guix system, I'm basically always using a fullscreen
terminal so haing an entire desktop environment is bloat. Switched it to
use kmscon so that I can still have pretty colors in my shell and vim.

: Set up /usr/run/$UID

This is the XDG_RUNTIME_DIR, which is assumed to exist in some cases,
for example during guix home setup. This is normally created by some
service that the desktop environment runs on, but as noted above this
has been removed from my system.

The current implementation is pretty hacky, requiring that UIDs and GIDs
are defined explicitly by the user. I think it is reasonable to expect
that these are available at declaration time, since the whole point of
guix is to statically declare everything so having the system "allocate
when the user is created" seems to be a violation of this. But I'm not
sure what the project maintainers would think about that.

: Make neovim's tabs useful

Add the patch which makes tab characters use an absolute value for their
width, instead of being calculated based on column position.

: Separate system packages from essential user packages

There are some packages that should always be present in the operating
system definition, and some of them have problems if installed as a user
package. For example, the sudo command does not work if installed as a
user package (and if it did then this would be a security
vulnerability). The essential-packages list has been split into
system-packages and essential-packages. The former *must* be included in
all operating-system definitions and must *not* be included in any user
definitions, while the latter may be included in either, both, or none.