Add haunt version 0.3.0
The Guix repository uses the release tarballs, this is building off of a
direct commit. This was released recently, so I'll give things a couple
of weeks to move and poke people if nothing seems to be happening. I
think it's better to let people do things in their own time unless there
is a specific expectation that I am doing something, but also it seems
rude to keep the updated package definition for myself and not share it
when it's readily available.