Update check phase to load dependency paths.
This also moves the `%load-path`-modifying code into the `check` wrapper
function (instead of `check-runner`). Currently, `check` doesn't do much
except call `check-wrapper`, which helped readability by avoidig nested
quasiquotes. However, it seems reasonable to have `check` be responsible
for environment setup/teardown, and have `check-wrapper` perform
the actual work. The thing that was "broke" was under-utilization of
`check`, which existed for a separate purpose before this commit.

The other thing that was "broke" was not loading dependent modules,
which is the main point of this commit, but doesn't need as much space
devoted to justification. =)