Add README to guix project
It is possible for free software projects to be successful, and when
they are they can create things better than anything the proprietary
world ever could. However, free software requires that developers are
responsible.

One kind of irresponsible thing which has historically happened multiple
times is that some person or group has decided to take a free software
project, rebrand it, and effectively take credit for the work. It does
not matter whether the people doing this were trying to act in good
faith ("customers don't care who wrote this code") or were actively
malicious: the effect is the same. For this reason, it is extra
important in free software communities to acknowledge the work of others
and be clear about what you are doing and what value you added vs what
value you borrowed.

When I first put together this repository, I did not think much about
these things because I am an obscure person and this is an obscure
repository. However, I have started linking to some file in my
repository in a few places, and if people following those links decide
to dig around I do not want them to have a bad impression of me.