Add os-fragments & integrate with existing code
The os-fragments module defines an API which makes it easier to define
incomplete operating-system definitions which are concerned with some
particular aspect of the system. Then, these incomplete definitions can
be included in a complete definition. This makes it easier to reason
about what is needed at any given point in system assembly. There are 2
main outstanding issues.

One is that there can be conflicts between fragments, for example if
they define exclusive services of the same type. Sometimes this conflict
represents a real conflict between fragments that needs to be resolved,
but sometimes they are compatible and fixing it is busywork.

The other is that the os-fragment API is deficient. This requires
enhancing `define-record-type*` from `guix records`. I still need to do
theoretical work on the modules which want this enhancement, and I would
prefer to develop that further before trying to update upstream code, in
case the further developments are relevant. See the documentation in
`collections.scm` for the theoretical portion and `os-fragment.scm` for
the deficiencies.