Add web module.
The web module is starting with 2 files, x509.scm and
examples/x509-echo.scm. The former provides some helper functions for
working with x509 certificates, primarly using gnutls as the backend but
reverting to openssl for certificate generation due to the author's
unfamiliarity with x509 internals.

The example file creates an echo server and a client which talks to it
over a TLS connection. A new authority certificate/key pair is created
and trusted by the client. A new server certificate/key pair is created,
signed by the authority, and used by the server.