Move code in root package.scm into guix package
I originally defined local packages in a root file that was never
installed to the system. This was to avoid a problem where the hash of
the source was dependent on the hash of the source. However, the build
strategy has significantly changed since then: now the source is
directly embedded into the build expression with local-file and source
is #f, so this problem no longer exists.