Run guile test modules in new guile processes.
For some reason, when I tried to bootstrap on a new machine the modules
in the r7rs directory were not read with the r7rs syntax (the `(srfi N)`
forms in the import directive were not recognized as valid).

I was never entirely comfortable with loading the package code directly
into the build process (seems likely to lead to security issues), so I
decided to just run the test in new guile processes as I should have the
whole time.