Use a per-invocation directory for signature output
The signed posts were previously all placed in the same directory,
under the reasoning that it was static data. However, from the
perspective of a user-agent, it is far more convenient for the source to
be placed in the same directory as the HTML. Additionally, this avoids
any potential problems if there are sources with the same filename in
different directories.