Make comment more technically accurate.
"Bottleneck" does not have the typical negative connotation here,
because making things pass through them in order to establish trust is
useful, while in performance optimization problems it is normally not
useful. However, "bottleneck" is more accurate than "root" because
they are not at the root, they are in the middle.