Delete guix code.
This code is being moved to a channel for a couple of reasons.

First, it makes it easier to bootstrap a new machine: instead of
downloading the repository and installing an ad-hoc package from a
local file, I can just add the channel (and it will automatically
pick up the guix-on-qubesos dependency).

Second, I keep getting local builds of packages when substitutes
really should be available. I suspect that this has to do with the
way I import guix modules from my system's runtime instead of using
the channel system as expected. From what I understand importing the
modules manually *shouldn't* cause any problems because the inputs
should be the same so it should be requesting the same hash from the
substitute server. But that's on paper and many things could
interfere with that expectation in practice. I don't have time to
chase down a bug that comes from using the API in an unexpected
manner when (if) using it in the expected manner makes everything
worke.